minor change and creation of dataharvester
Trying to do datharverster that should tkae the data and save it as excel files for each electrode with information by muscle and current about p2p and rsponse. This will be used for another file that will be created where we load these csv to make the electrode classification.